bin/stream_controller/requirements.txt
playwright
beautifulsoup4
pyserial-asyncio
requests
